==> lib/01_map_dfg2_to_wweia.ipynb <==
### Map DFG2 foods to WWEIA ingredient codes

import pandas as pd

#load data for the four DFG2 glycan analyses and wweia map
poly = pd.read_csv('../../data/glycan/polysaccharide_v2.csv')
free = pd.read_csv('../../data/glycan/free_saccharide_v2.csv')
free.drop(columns={'simple_name','food_group'},inplace=True)
mono = pd.read_csv('../../data/glycan/monosaccharide_v2.csv')
mono.drop(columns={'simple_name','food_group'},inplace=True)
linkage = pd.read_csv('../../data/glycan/glycosidic_linkage_v2.csv')
linkage.drop(columns={'simple_name','food_group'},inplace=True)
wweia_map = pd.read_csv('../../data/wweia/dfg2_wweia_map.csv')

free.set_index('sample_id', inplace=True)
free.columns = ['Free_' + str(col) for col in free.columns]
free.reset_index(inplace=True)

mono.set_index('sample_id', inplace=True)
mono.columns = ['Mono_' + str(col) for col in mono.columns]
mono.reset_index(inplace=True)

poly_free = poly.merge(free, on='sample_id', how='left')
poly_free_mono = poly_free.merge(mono, on='sample_id', how='left')
dfg_all = poly_free_mono.merge(linkage, on='sample_id', how='left')

wweia_map = wweia_map[['sample_id', 'ingred_code']]
wweia_map.dropna(inplace=True)

dfg_wweia = wweia_map.merge(dfg_all, on='sample_id', how='left')

dfg_wweia.to_csv('../../data/wweia/dfg2_wweia_data.csv',index=None)

### calculate absolute abundance of linkage feature (rel. abun.) by multiplying by monosaccharide values

linkage.iloc[:,1:10] = linkage.iloc[:,1:10].mul(mono['Mono_Glucose'], axis = 0, level = 0) #glucose

linkage.iloc[:,10:17] = linkage.iloc[:,10:17].mul(mono['Mono_Galactose'], axis = 0, level = 0) #galactose

linkage.iloc[:,17:23] = linkage.iloc[:,17:23].mul(mono['Mono_Xylose'], axis = 0, level = 0) #xylose

linkage.iloc[:,23:29] = linkage.iloc[:,23:29].mul(mono['Mono_Arabinose'], axis = 0, level = 0) #arabinose

linkage.iloc[:,29] = linkage.iloc[:,29] * mono.Mono_Fucose #fucose

linkage.iloc[:,30:34] = linkage.iloc[:,30:34].mul(mono['Mono_Rhamnose'], axis = 0, level = 0) #rhammose

linkage.iloc[:,36:43] = linkage.iloc[:,36:43].mul(mono['Mono_Mannose'], axis = 0, level = 0)  #mannose

#linkage = linkage.drop(columns=['4-Glucosamine/GlcNac', '3-Glucosamine/GlcNac']) 

linkage = linkage.iloc[:,0:41]

linkage
