==> lib/05a_train_test_split.ipynb <==
import pandas as pd

x_in = pd.read_csv('../output/04/mod_filt.csv', index_col=['SEQN'])

y_in = pd.read_csv('../output/02/outcome_homa.csv', index_col=['SEQN'])

sw = pd.read_csv('../output/02/sample_wt_homa.csv', index_col=['SEQN'])

print(x_in.shape)
print(y_in.shape)
print(sw.shape)

x_in_dummy = pd.get_dummies(x_in, dtype=float)

from sklearn.model_selection import train_test_split

y_train, y_test, x_train, x_test, sw_train, sw_test = train_test_split(y_in, x_in_dummy, sw, stratify=y_in['ir'], shuffle=True, test_size=0.2)

print(x_train.shape)
print(x_test.shape)
print(y_train.shape)
print(y_test.shape)

x_train.columns

y_train.to_csv('../output/05/y_train.csv', index=True)
x_train.to_csv('../output/05/x_train.csv', index=True)
y_test.to_csv('../output/05/y_test.csv', index=True)
x_test.to_csv('../output/05/x_test.csv', index=True)
sw_train.to_csv('../output/05/sw_train.csv', index=True)
