==> lib/05b_train_test_split_no_alcohol.ipynb <==
import pandas as pd

no_alc = pd.read_csv('../output/02/dfg2_homa_no_alc.csv')

x_in = pd.read_csv('../output/04/mod_filt.csv')

y_in = pd.read_csv('../output/02/homa_outcome_no_alc.csv', index_col=['SEQN'])

sw = pd.read_csv('../output/02/sample_wt_homa.csv')

x_in = x_in[x_in['SEQN'].isin(no_alc['SEQN'])]

x_in.shape

sw = sw[sw['SEQN'].isin(x_in['SEQN'])]

sw = sw.set_index('SEQN')

x_in = x_in.set_index('SEQN')

x_in.columns

print(x_in.shape)
print(y_in.shape)
print(sw.shape)

x_in_dummy = pd.get_dummies(x_in, dtype=float)

from sklearn.model_selection import train_test_split

y_train, y_test, x_train, x_test, sw_train, sw_test = train_test_split(y_in, x_in_dummy, sw, stratify=y_in['ir'], shuffle=True, test_size=0.2)

print(x_train.shape)
print(x_test.shape)
print(y_train.shape)
print(y_test.shape)

y_train.to_csv('../output/05/y_train_no_alc.csv', index=True)
x_train.to_csv('../output/05/x_train_no_alc.csv', index=True)
y_test.to_csv('../output/05/y_test_no_alc.csv', index=True)
x_test.to_csv('../output/05/x_test_no_alc.csv', index=True)
sw_train.to_csv('../output/05/sw_train_no_alc.csv', index=True)
